Clean up hit test explainer & add use cases
The change adds a variant of the API that allows to pre-subscribe to
transient input sources based on their XRTargetRayMode. This should be a
reasonable compromise that does not require the removal of transient input
sources concept from the main spec, but still allows the application to
specify arbitrary rays for hit tests on transient input sources and does
not force implementations to speculatively perform hit test for newly
created transient input sources in hopes that their results will be
requested by the application. Additionally, we can consider adding more
options to filter out transient input sources (ex. by their XRHandedness or
by the profile that they support).

This removes asynchronous variant of hit test API as there seems to be no
clarity on how to expose its results. The current API should be extensible
so that once we agree on the shape of (and the need for) asynchronous API,
we should be able to easily add it back.

The change also removes Grab&drag section as based on the code there, it
was relevant exclusively for virtual hit test (the hit test result *had* to
have `virtualTarget` entry) and did not depend in any way on the proposed
hit test API.

There are some other minor changes to the proposed IDL, namely:
- XRHitTestResult.getPose() XRSpace parameter is no longer optional & got
renamed to baseSpace to match XRFrame.getPose()
- XRHitTestOptions interface & its uses got removed as the convention is
to not put the objects that were required to create some instances as the
attributes on those instances - the application can attach them if needed
- async variant of the API was removed (as mentioned)

Modify API to use more generic input profile name for TIS filtering